+++ src/media_enrichment/__init__.py
@@ -1,7 +1,7 @@
 """media-enrichment Skill package.
 
-Version: 0.1.0-dev7-hotfix3
+Version: 0.1.0-dev7-hotfix4
 """
 
-__version__ = "0.1.0-dev7-hotfix3"
+__version__ = "0.1.0-dev7-hotfix4"
 SKILL_NAME = "media-enrichment"